requirements.txt
aioredis==1.3.1
asgiref==3.2.10
async-timeout==3.0.1
attrs==20.1.0
autobahn==20.7.1
Automat==20.2.0
autopep8==1.5.4
cffi==1.14.2
channels==2.4.0
channels-redis==3.0.1
constantly==15.1.0
cryptography==3.0
daphne==2.5.0
Django==3.1
djangorestframework==3.11.1
hiredis==1.1.0
hyperlink==20.0.1
idna==2.10
importlib-metadata==1.7.0
incremental==17.5.0
jsonschema==3.2.0
msgpack==1.0.0
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
pyrsistent==0.16.0
pytz==2020.1
service-identity==18.1.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
Twisted==20.3.0
txaio==20.4.1
zipp==3.1.0
zope.interface==5.1.0
